Splitting up action and adding rules tests
