+++ packages/action/jest.config.js
@@ -1,0 +1,3 @@
+const config = require('../../jest.config');
+
+module.exports = config;
