+++ packages/action/src/types.ts
@@ -1,0 +1,3 @@
+import { getOctokit } from '@actions/github';
+
+export type OctokitInstance = ReturnType<typeof getOctokit>;
